requirements.txt
-i https://pypi.org/simple
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
djongo>=1.3.0
dnspython==2.2.1
django==4.0.4
django-cors-headers==3.11.0
django-crontab==0.7.1
djangorestframework==3.13.1
idna==3.3
pymongo==3.12.1
pytz==2022.1
requests==2.27.1
sqlparse==0.2.4
urllib3==1.26.9
